+++ src/components/addBook/AddBookWrapper.tsx
@@ -6,25 +6,11 @@ import { MainHeader, MainLayout } from "../common";
 import BookList from "./BookList";
 import SearchBar from "./SearchBar";
 
-export interface BookProps {
-  thumbnail: string;
-  title: string;
-  authors: string[];
-  datetime: Date;
-  contents: string;
-}
 export default function AddBookWrapper() {
   const headerColor = theme.colors.orange100;
 
   const [books, setBooks] = useState<any[]>([]);
   const [query, setQuery] = useState("");
-  const [bookInfo, setBookInfo] = useState<BookProps>({
-    thumbnail: "",
-    title: "",
-    authors: [],
-    datetime: new Date(),
-    contents: "",
-  });
 
   const bookSearchHandler = async (query: string, reset: boolean) => {
     const params = {
@@ -46,21 +32,6 @@ export default function AddBookWrapper() {
   useEffect(() => {
     if (query.length > 0) {
       bookSearchHandler(query, true); // 컴포넌트 마운트 후에, 함수를 호출한다.
-      if (books.length > 0) {
-        const book1 = books[0];
-
-        const { thumbnail, title, authors, datetime, contents }: BookProps = book1;
-
-        setBookInfo((prev: any) => ({
-          ...prev,
-          thumbnail,
-          title,
-          authors,
-          datetime,
-          contents,
-        }));
-        console.log(bookInfo);
-      }
     }
   }, [query]);
 
@@ -68,7 +39,7 @@ export default function AddBookWrapper() {
     <MainLayout>
       <MainHeader color={headerColor}>책 추가</MainHeader>
       <SearchBar setQuery={setQuery} />
-      <BookList bookInfo={bookInfo} />
+      <BookList books={books} />
     </MainLayout>
   );
 }
